FIX: Allow configuration of IPMI privilege level
This change updates the ansible tasks to supply a privilege level (configured as
hwmgmt.priv_level) when calling ipmitool.  Without this change, ipmitool defaults
to using the 'ADMINISTRATOR' privilege.  This will fail on systems where
hwmgmt.user is assigned a lower privilege level.

This change depends on the following changes:
https://gerrit.akraino.org/r/c/ta/config-manager/+/1076
https://gerrit.akraino.org/r/c/ta/hw-detector/+/1077
https://gerrit.akraino.org/r/c/ta/cm-plugins/+/1078

signed-off-by: dave kormann <davek@research.att.com>
Change-Id: If97a52fa2f2dbf903f2c8e2311e163b2ffd1b944