FIX: Change ownership of newly-created VGs
When creating a new volume group, allow a user/group to be specified
to own the root directory.  If no user/group is provided, use the
owner of the mountpoint or (if it doesn't exist) the current user/group.

This should fix the problem with mariadb permission problems when
installing to a vg, as opposed to the existing directory in the
image.

signed-off-by: dave korman   <davek@research.att.com>
Change-Id: Idf4418dca288a374c53fa6c180c3bdf62db10abb