Update patch set 5
Patch Set 5:

Removed the Akraino Jobbuilder -1 because it is not due to this change. It is probably due to the builder nodes changing from Ubuntu 16.04 to 18.04 and there is a separate change in the works to try to fix that.

Patch-set: 5
Reviewer: Gerrit User 43 <43@4cada01b-0314-4fe0-84a1-f9ddc89863f9>